feat(ExamController,IExamService,Difficulty,QuestionSubmissionRepo,SubmitExamRequest,SubmitMCQQuestionRequest,SubmitProgrammingQuestionRequest,ExamService) : submitExam, Calculate marks feature added

+++ src/main/java/com/oea/online_exam_app/Models/Difficulty.java
@@ -24,7 +24,7 @@ public class Difficulty {
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     private int difficultyId;
     private String difficultyText;
-    private float difficultyWeight;
+    private int difficultyWeight;
 
     public Difficulty(){}
